Add a function to build a default page template, and
also add sample files to try out the process. The page builder can use
additional 'page_data' context, if it is set by the caller. We try to
handle optional page_data gracefully.

The function makes no assumption about the type of content file. We
combine this "dumbness" with a mandatory method to translate the file
to HTML. This way we can author content in nearly any format we are
comfortable with, and use any convenient tool to translate it to HTML.