Define a convention around HTML files for content authoring
Suppose the first line of a page is an HTML comment, having page data?

Now suppose we literally write down an associative array in Bash?

Like so:

  <!-- ([page_id]="hello-world-page") -->
  <h1>Hello, world!</h1>
  <p>How are you doing today?</p>
  <p>I'm here.</p>
  <p>And I'm going to take you head-on...</p>

Suppose you stop shivering with fear, and just play along for a bit?
:D

Now we can do something like this in our shell session:

  $ declare -A page_data="$(get_html_page_data ./sample/hello-data.html)"

  $ shite_build_page /sample/hello-data.html except_html_page_data

Notice that the page_id we declared in hello-data.html gets injected
into the page. Rejoice a little!