Add function to build multiple HTML pages and write them
to the 'public' html directory. Also add more detailed content for a
website that is aligned with the <nav> menu items we have been
carrying around so far.

At this point we have a functioning website with pages for "About",
"Resume", and a "Blog" that doubles as the home page. It is
spartan. No CSS or style yet.