Create entry point for hotreloaded shite making
Refactor the templating code so that we have global access to the
array holding global site data. This will allows the new shite entry
point to either override or fall back to the baked-in default values.
At some point we may decide to pull this data out into a JSON config
file or something.