Remove global state from business logic. Explicitly pass
as parameters. I exercised the hotreload workflow WITHOUT setting the
global page array and noticed this error in the console, which
suggested something was awry somewhere with parameter expansion:

     -bash: file:///path/to/shite: syntax error in expression (error
      token is ":///path/to/shite")

It's subtle. The _array has to be entirely unset_, not merely missing
the 'canonical_url' key/value pair.

The bad substitution error revealed another gotcha. When it failed, it
caused 'read -r key val' to set key and val as empties, which in turn
caused a 'bad array subscript' error (observed with log statements):

  ++ __to_stderr 'INFO -bash KEY: , VAL: '
  ++ printf '%s\n' '2022-09-02T17:03:07+05:30 INFO -bash KEY: , VAL:
  2022-09-02T17:03:07+05:30 INFO -bash KEY: , VAL:
  ++ shite_page_data[${key}]=
  -bash: shite_page_data[${key}]: bad array subscript

So, I also threw in a guard against empty or blank lines in the
combined stream of parsed page front matter plus extra injected page
metadata. If blank lines occur, we get key error when setting the page
data array.

Below is the relevant part of the full trace, which shows the chain of
substitution error to array error. Debugged with 'set -xvT'.

  ++ __shite_hot_cmd_browser_refresh 14680124
  ++ local window_id=14680124
  ++ printf '%s\n' 'key --window 14680124 --clearmodifiers '\''F5'\'''
  ++ prev_file_name=index.org
  ++ IFS=,
  ++ read -r timestamp event_type watch_dir url_slug file_name
     file_type content_type key --window 14680124 --clearmodifiers 'F5'
  -bash: file:///path/to/shite: syntax error in expression (error
         token is ":///path/to/shite")
  +++ cat /dev/fd/63
  ++ local optional_metadata_csv=
  ++ IFS=,
  ++ read -r key val
  +++ cat /path/to/shite/sources/posts/hello/index.org
  +++ cat -
  +++ __shite_templating_get_page_front_matter org
  +++ local file_type=org
  +++ case ${file_type} in
  +++ stdbuf -oL grep -v '^\s?$'
  +++ sed -n -E -e '/^\#\s+shite_meta/I,/^\#\s+shite_meta/I{/\#\s+shite_meta.*/Id; s/^\#\+(\w+)\:\s+(.*)/\L\1\E,\2/Ip}'
  ++ __log_info 'KEY: , VAL: '
  +++ echo 'INFO -bash KEY: , VAL: '
  ++ __to_stderr 'INFO -bash KEY: , VAL: '
  ++ [[ debug == \d\e\b\u\g ]]
  +++ date --iso-8601=seconds
  ++ printf '%s\n' '2022-09-02T17:03:07+05:30 INFO -bash KEY: , VAL: '
  2022-09-02T17:03:07+05:30 INFO -bash KEY: , VAL:
  ++ shite_page_data[${key}]=
  -bash: shite_page_data[${key}]: bad array subscript