Fix shite_publish case match patterns. They were all wrong
due to my misunderstanding of how "Pattern Matching" worked in the
case clauses. cf. the Bash manpage sections on case / esac and for
Pattern Matching. In short, here is the wrong/right way to use `|` to
specify alternatives.

WRONG way; try to specify optional parts of a match term:

  case ${x}:${y} in
    foo|bar:* )
    echo "This won't echo."
    ;;
    * )
    echo "Catchall will always echo."
    ;;
  esac

RIGHT way; specify options for _entire_ terms:

  case ${x}:${y} in
    foo:*|bar:* )
    echo "This will echo when ${x} is either 'foo' OR 'bar'"
    ;;
    * )
    echo "Catchall will echo IFF ${x} is NEITHER 'foo' NOR 'bar'."
    ;;
  esac