Add handy trick to modify templates, without restartinga
our pipeline each time we change template functions.

This still requires us to modify/touch content that gets templated,
in order to cause the page to re-render on-screen, with the template
modifications.

Ideally the current page should re-render if I modify a template. But
the code for that may be too subtle for anyone's good.

Also having to modify a content file is a fine tradeoff, as presumably
I'll be interested in previewing the effect of a template change
against the content in question.