Refactor hotreload fn. Lift out details of events and
of templating into their respective namespaces. Also update the name
of the main publishing function to better reflect its use. And remove
dependence on title in the browser tab as that makes the browser
interactions brittle, and causes annoying race conditions if a title
is changed or not present v/s whatever is configured in global data.