Fix generation and injection of ToC into a post's
body. We want a well-defined automatic Table of Contents in our
posts. Pandoc generates ToC only if we use the -s flag to generate a
full HTML page. This is unpalatable because now we have to parse out
blocks that we want to stuff into our templates. However it may not be
that bad, because ToC generation is more reliable than parsing out <h>
tags and translating to ToC ourselves.

We also use regex to parse HTML and suffer. sed's regex is greedy. So
our regex was mis-parsing headings that contained HTML elements such
as <em>...</em> or <strong>...</strong>.

The new regex assumes a very well defined pandoc-generated "vanilla"
heading style:

  <h2 id="some-id">Some text <em>emphasised text</em>.</h2>

If headings conform to this expectation, all the text (html included)
between the open/close `h` tags gets captured in the interposed (.*)
capture group (greedily).

Templating, oh templating. Sigh.