Don't get clever with mkfifo. As Prof. McIlroy admonished,
"stdio's buffering is evil!".

It turns out that the hotbuild pipeline will choke at the named pipe,
for a sufficiently large article, such as the post on "n ways to
FizzBuzz in Clojure". I suppose it has something to do with stdio
buffering. Besides, a blocking pipe is the wrong abstraction for this
template generation job, at the outset.