Add batch mode to rebuild the whole site. Not the best
way to do it, but we need to ship now. I think this is showing the
fact that the semantics of events is too tightly coupled with the
publishing logic.

Anyway, I've modified the "debug" flag to a "hotreload" flag to make
the notion of "not-hotreload" (i.e. batch mode) more explicit.

We use inotifywait's timeout facility to cobble together "batch mode"
behaviour. A suitably small timeout ensures it terminates soon enough,
if the watched files remain untouched. The trick is to start the usual
hotreload behaviour but background it so that we can immediately
trigger file events on sources, that will in turn cause hotreload to
publish. We can do this simply by updating mtime on all the source
files. Care must be taken to suppress xdotool / browser hotreload
command execution.