Add Cloudflare analytics. Other privacy-respecting
analytics providers are not affordable at this time. Once a commercial
case can be made for content published (if it can be made), then the
other options like fathom may become viable.

The purpose of analytics is educational for now. If monkey-brain starts
getting hooked, then I will probably get rid of it.